redis service error handler

+++ src/services/RedisService.ts
@@ -2,61 +2,34 @@
 import Redis from "ioredis";
 import { AssetData, Exchange, WSChannel } from "../config/enums";
 import { IRemoveSignal, ISignalOrderBook, ISignalPrice } from "../config/interfaces";
-import * as WebSocketType from "ws";
 
 export class RedisClient {
-	private readonly client: Redis;
+	private client: Redis | null = null;
 	private readonly env: string;
 
 	constructor() {
+		this.env = process.env.NODE_ENV as string;
 		this.client = new Redis({
 			host: process.env.REDIS_URL,
 			port: Number(process.env.REDIS_PORT),
+			maxRetriesPerRequest: 5, // Stops reconnectiong after 5 failed attempts
 		});
-		this.env = process.env.NODE_ENV as string;
+
+		this.client.on("ready", () => console.log("Redis connection established"));
+
+		// Handles error in redis connection
+		this.client.on("error", (err) => {
+			console.error("Redis connection error:", err);
+			this.client = null; // Mark redis as unavailable
+		});
 	}
-
-	/*
-    -------------------
-    ACTIVE USERS
-    -------------------
-    */
-
-	async addChannelClient({ userId, ws }: { userId: string; ws: string }): Promise<void> {
-		const wsKey = `${this.env}_${WSChannel.usersWs}_${userId}`;
-		this.client.set(wsKey, ws);
-	}
-
-	async getChannelClients(): Promise<WebSocketType[]> {
-		const keys = await this.client.keys(`${this.env}_${WSChannel.usersWs}_*`);
-		const clients = await Promise.all(
-			keys.map(async (key) => {
-				const client = await this.client.get(key);
-				return JSON.parse(client ?? "") as WebSocketType;
-			})
-		);
-
-		return clients;
-	}
-
-	async getChannelClient(userId: string): Promise<WebSocket | null> {
-		const wsKey = `${this.env}_${WSChannel.usersWs}_${userId}`;
-		const client = await this.client.get(wsKey);
-		return client ? JSON.parse(client ?? "") : null;
-	}
-
-	async removeChannelClient(userId: string): Promise<void> {
-		const wsKey = `${this.env}_${WSChannel.usersWs}_${userId}`;
-		this.client.del(wsKey);
-	}
-
-	/*
-    -------------------
-    ACTIVE SIGNALS
-    -------------------
-    */
 
 	async addSignalPrice({ signalId, exchange, asset, assetPrice }: ISignalPrice) {
+		if (!this.client) {
+			console.warn("Redis is unavailable. Skipping operation.");
+			return;
+		}
+
 		const wsKey = `${this.env}_${WSChannel.assetsUpdateWs}_${AssetData.price}_${signalId}_${exchange}`;
 		const data = {
 			asset,
@@ -71,15 +44,26 @@ export class RedisClient {
 		totalSellQuantityInRange,
 		totalBuyQuantityInRange,
 	}: ISignalOrderBook) {
+		if (!this.client) {
+			console.warn("Redis is unavailable. Skipping operation.");
+			return;
+		}
+
 		const wsKey = `${this.env}_${WSChannel.assetsUpdateWs}_${AssetData.orderBook}_${signalId}_${exchange}`;
 		const data = {
 			totalBuyQuantityInRange,
 			totalSellQuantityInRange,
 		};
+
 		await this.client.set(wsKey, JSON.stringify(data));
 	}
 
 	async getAllSignalsPrices(exchange?: string): Promise<ISignalPrice[]> {
+		if (!this.client) {
+			console.warn("Redis is unavailable. Skipping operation.");
+			return [];
+		}
+
 		const filteredKey = exchange
 			? `${this.env}_${WSChannel.assetsUpdateWs}_${AssetData.price}_*_${exchange}`
 			: `${this.env}_${WSChannel.assetsUpdateWs}_${AssetData.price}_*`;
@@ -90,7 +74,7 @@ export class RedisClient {
 				const keyArray = key.split("_");
 				const signalId = keyArray[keyArray.length - 2];
 				const exchange = keyArray[keyArray.length - 1] as Exchange;
-				const assetValue = (await this.client.get(key)) as string;
+				const assetValue = (await this.client?.get(key)) as string;
 				const { asset, assetPrice } = JSON.parse(assetValue);
 
 				return { signalId, exchange, asset, assetPrice };
@@ -101,6 +85,11 @@ export class RedisClient {
 	}
 
 	async getAllSignalsOrderBooks(exchange?: string): Promise<ISignalOrderBook[]> {
+		if (!this.client) {
+			console.warn("Redis is unavailable. Skipping operation.");
+			return [];
+		}
+
 		const filteredKey = exchange
 			? `${this.env}_${WSChannel.assetsUpdateWs}_${AssetData.orderBook}_*_${exchange}`
 			: `${this.env}_${WSChannel.assetsUpdateWs}_${AssetData.orderBook}_*`;
@@ -111,7 +100,7 @@ export class RedisClient {
 				const keyArray = key.split("_");
 				const signalId = keyArray[keyArray.length - 2];
 				const exchange = keyArray[keyArray.length - 1] as Exchange;
-				const assetValue = (await this.client.get(key)) as string;
+				const assetValue = (await this.client?.get(key)) as string;
 				const { totalBuyQuantityInRange, totalSellQuantityInRange } =
 					JSON.parse(assetValue);
 				return { signalId, exchange, totalSellQuantityInRange, totalBuyQuantityInRange };
@@ -122,29 +111,49 @@ export class RedisClient {
 	}
 
 	async removeSignalPrice({ signalId, exchange }: IRemoveSignal): Promise<void> {
+		if (!this.client) {
+			console.warn("Redis is unavailable. Skipping operation.");
+			return;
+		}
 		const wsKey = `${this.env}_${WSChannel.assetsUpdateWs}_${AssetData.price}_${signalId}_${exchange}`;
 		await this.client.del(wsKey);
 	}
 
 	async removeSignalOrderBook({ signalId, exchange }: IRemoveSignal): Promise<void> {
+		if (!this.client) {
+			console.warn("Redis is unavailable. Skipping operation.");
+			return;
+		}
 		const wsKey = `${this.env}_${WSChannel.assetsUpdateWs}_${AssetData.orderBook}_${signalId}_${exchange}`;
 		await this.client.del(wsKey);
 	}
 
-	async deleteAllCacheRecord() {
-		const priceFilteredKey = `${this.env}_${WSChannel.assetsUpdateWs}_${AssetData.price}_*`;
-		const orderBookFilteredKey = `${this.env}_${WSChannel.assetsUpdateWs}_${AssetData.orderBook}_*`;
+	async deleteAllCacheRecord(): Promise<void> {
+		if (!this.client) {
+			console.warn("Redis is unavailable. Skipping operation.");
+			return;
+		}
 
-		const priceKeys = await this.client.keys(priceFilteredKey);
-		const orderBookKeys = await this.client.keys(orderBookFilteredKey);
+		const key = `${this.env}_${WSChannel.assetsUpdateWs}_*`;
 
-		const delPrice = priceKeys.map((price) => this.client.del(price));
-		const delOrderBook = orderBookKeys.map((order) => this.client.del(order));
+		const keys = await this.client.keys(key);
 
-		await Promise.all([...delPrice, ...delOrderBook]);
+		if (keys.length > 0) {
+			await this.client.unlink(...keys);
+		}
+
+		await this.closeConnection();
 	}
 
-	closeConnection(): void {
-		this.client.disconnect();
+	async closeConnection(): Promise<void> {
+		if (!this.client) {
+			console.warn("Redis is unavailable. Skipping operation.");
+			return;
+		}
+		try {
+			await this.client.quit();
+		} catch (error) {
+			console.error("Error closing Redis connection:", error);
+		}
 	}
 }
